requirements.txt
streamlit
tensorflow
numpy
pandas
Pillow
joblib
openpyxl